wrote action creator and thunk creator for create single meme. Created forms folder under components to hold form components, and moved AuthForm to said folder.
