+++ client/components/forms/AuthForm.js
@@ -1,21 +1,21 @@
-import React from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import {authenticate} from '../store'
+import React from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { authenticate } from '../../store';
 
 /**
  * COMPONENT
  */
 const AuthForm = ({ name, displayName }) => {
-  const { error } = useSelector(state => state.auth)
-  const dispatch = useDispatch()
+  const { error } = useSelector((state) => state.auth);
+  const dispatch = useDispatch();
 
   const handleSubmit = (evt) => {
-    evt.preventDefault()
-    const formName = evt.target.name
-    const username = evt.target.username.value
-    const password = evt.target.password.value
-    dispatch(authenticate(username, password, formName))
-  }
+    evt.preventDefault();
+    const formName = evt.target.name;
+    const username = evt.target.username.value;
+    const password = evt.target.password.value;
+    dispatch(authenticate(username, password, formName));
+  };
 
   return (
     <div>
@@ -38,8 +38,8 @@ const AuthForm = ({ name, displayName }) => {
         {error && error.response && <div> {error.response.data} </div>}
       </form>
     </div>
-  )
-}
+  );
+};
 
-export const Login = <AuthForm name="login" displayName="Login" />
-export const Signup = <AuthForm name="signup" displayName="Sign Up" />
+export const Login = <AuthForm name="login" displayName="Login" />;
+export const Signup = <AuthForm name="signup" displayName="Sign Up" />;
