+++ client/store/allMemes.js
@@ -1,10 +1,16 @@
 import axios from 'axios';
 
 const SET_MEMES = 'SET_MEMES';
+const CREATE_MEME = 'CREATE_MEME';
 
 const setMemes = (memes) => ({
   type: SET_MEMES,
   memes,
+});
+
+const newMeme = (meme) => ({
+  type: CREATE_MEME,
+  meme,
 });
 
 export const getMemes = () => {
@@ -18,10 +24,23 @@ export const getMemes = () => {
   };
 };
 
+export const createMeme = (meme) => {
+  return async (dispatch) => {
+    try {
+      const { data: created } = await axios.post('/api/memes', meme);
+      dispatch(newMeme(created));
+    } catch (error) {
+      console.error(error);
+    }
+  };
+};
+
 const memesReducer = (memes = [], action) => {
   switch (action.type) {
     case SET_MEMES:
       return action.memes;
+    case CREATE_MEME:
+      return [...memes, action.meme];
     default:
       return memes;
   }
